pyproject.toml
[tool.poetry]
name = "disc-golf-stats"
version = "0.1.0"
description = ""
authors = ["Anders Robstad <anders.robstad@live.no>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
whitenoise = "^5.2.0"
gunicorn = "^20.1.0"
django-heroku = "^0.3.1"
djangorestframework-simplejwt = "^4.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
